fixed path links in moved files
