+++ configAutoencoders.yml
@@ -33,3 +33,4 @@ config6:
   batchSize: 8
   epochs: 30
   network: 'leCunhaNet'
+
