+++ scripts/shared/autoencoderHelpers.py
